fix(livepeer): copy the missing files and add review comments
created layout for reusing the livepeer client code
