+++ .env.example
@@ -10,6 +10,8 @@ NEXT_PUBLIC_ALCHEMY_API_KEYs=
 # Infura: https://www.infura.io
 NEXT_PUBLIC_INFURA_API_KEY=
 
+NEXT_PUBLIC_LIVEPEER_API_KEY="481df680-dda5-470a-af77-f1a66298d557"
+
 # -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 # Private Variables
 # -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
