+++ package.json
@@ -35,6 +35,7 @@
   "packageManager": "pnpm@7.1.8",
   "dependencies": {
     "@lit-protocol/lit-node-client": "2.1.161",
+    "@livepeer/react": "^2.4.4",
     "@next/bundle-analyzer": "^13.1.1",
     "@next/font": "13.1.1",
     "@prisma/client": "^4.8.1",
